+++ package.json
@@ -4,8 +4,9 @@
   "description": "API to provide fake data for dummy data generation",
   "main": "server.ts",
   "scripts": {
-    "start": "node server.ts",
-    "watch": "nodemon server.ts"
+    "start": "vercel dev",
+    "watch": "nodemon server.ts",
+    "deploy" : "vercel deploy --prod"
   },
   "keywords": [
     "faker",
@@ -22,7 +23,8 @@
     "express": "^4.17.2",
     "faker": "^5.5.3",
     "ts-node": "^10.4.0",
-    "typescript": "^4.5.4"
+    "typescript": "^4.5.4",
+    "vercel": "^24.0.0"
   },
   "devDependencies": {
     "@types/express": "^4.17.13",
